Integer attributes handle unsigned ints

+++ src/main/java/org/tinyradius/attribute/AttributeType.java
@@ -119,7 +119,7 @@ public class AttributeType {
      * is an enumeration, or null if it is not or if the integer value
      * is unknown.
      */
-    public String getEnumeration(int value) {
+    public String getEnumeration(long value) {
         return enumeration.get(value);
     }
